requirements.txt
langchain
python-dotenv
langchain_community
langchain-openai
streamlit